[ refactor ] rename post package to board package
