+++ src/main/java/greeny/backend/domain/board/controller/PostController.java
@@ -1,9 +1,9 @@
-package greeny.backend.domain.post.controller;
+package greeny.backend.domain.board.controller;
 
 import greeny.backend.domain.member.service.MemberService;
-import greeny.backend.domain.post.dto.UpdatePostRequestDto;
-import greeny.backend.domain.post.dto.CreatePostRequestDto;
-import greeny.backend.domain.post.service.PostService;
+import greeny.backend.domain.board.dto.UpdatePostRequestDto;
+import greeny.backend.domain.board.dto.CreatePostRequestDto;
+import greeny.backend.domain.board.service.PostService;
 import greeny.backend.response.Response;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
