+++ src/main/java/greeny/backend/domain/board/dto/UpdatePostRequestDto.java
@@ -1,11 +1,10 @@
-package greeny.backend.domain.post.dto;
+package greeny.backend.domain.board.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.Max;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
