+++ src/main/java/greeny/backend/domain/board/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package greeny.backend.domain.post.repository;
+package greeny.backend.domain.board.repository;
 
-import greeny.backend.domain.post.entity.Post;
+import greeny.backend.domain.board.entity.Post;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
